+++ CHANGELOG.md
@@ -26,3 +26,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- Adds support for layout-07 (Limit line length to 115 characters)